Fix crash with Igor Pro 8 and long function names
The following results currently in a crash:
- Use the current XOP without long name support
- Run it under Igor Pro 8
- Try calling a function with a long name

When calling a function we call in CallFunctionOperation::CanBeProcessed()
from MessageHandler::CallAndReply() the XOP Toolkit function
GetFunctionInfo.

This function returns with an return value which later on triggers an
assertion.

This assertion is then propagated via an IgorException to CallAndReply.
As we currently use json::parse to read the exception message, this in
turn triggers an assertion from the json library as the exception
message holds some characters which need escaping for json.

And then finally the message handler thread crashes as we don't catch
that assertion.

For fixing that we move the catch(...) one level up.