Avoid code duplication in RequestInterface handling
We currently have two places where we call Call() from RequestInterface.
One is CallAndReply from the message handler thread and the other is
CallIgorFunctionFromMessage which is used from zmq_test_callfunction.

The code in CallIgorFunctionFromMessage does also misses setting the
messageID on error.

By introducing CallIgorFunctionFromReqInterface we can avoid code
duplication and always have a messageID in the error case.