IgorException: Avoid emergency output mode for normal errors
Since a691ef45 (Rework debug output functions, 2021-03-25) instances of
IgorException are outputted with EMERGENCY_OUTPUT.

This is not correct, as this is a plain error which can happen and we
should not alert the user in this case.

To handle that properly a new OutputMode termed Normal was implemented.

That does not use a header with function name, line and prefix but just
output everything.