src: Use correct bind/connect methods for Pub/Sub
We need to use zmq_connect/zmq_disconnect for Subscriber and
zmq_bind/zmq_unbind for Publisher.

This was not caught in our tests because somehow this works when done in
the same process nevertheless.

Bug introduced in 768c8cb9 (Add Publisher/Subscriber functionality, 2021-04-05).

This also finally explains the asymmetry between server/client sockets
vs publisher/subscriber sockets as mentioned in f2d4eff9 (Tests: Extend
existing tests to also cover pub/sub, 2021-04-06). There it was observed
that the publisher socket did not bind to a port as reported by netstat.