Upgrade fmt
The new version is the same as the one we use for the ITCXOP2 and has
support for outputting std::chrono objects which we want to use in a
future commit.

The new fmt version requires a std::back_inserter for fmt::format_to and
also the logging template had to be updated.

As we don't use fmt literals anymore, and they are deprecated as well,
that was also removed.