Logging::Implementation::AddLogEntry: Add a timestamp
In one of the ZeroMQ users [1] we were feeding in the timestamps via the
logging template. Which of course makes no sense as we want to have
the timestamp when adding the log entry not when setting up the logging.

Fix that.

We use the same approach as in the ITCXOP2.

[1]: https://github.com/AllenInstitute/MIES
[2]: https://github.com/AllenInstitute/ITCXOP2@ff863665 (LoggingImplementation::AddLogEntry: Add ISO8601 timestamp, 2022-09-26)