[UNTESTED] feat: update to support Igor Pro 6
Making changes to the code to support Igor Pro 6 (the Igor Pro version
used by Asylum Research). The existing version of ZeroMQ-XOP supports
Igor versions > 8, presumably because it was compiled with the XOP
Toolkit 8 (and had no reason to support earlier versions).

Unfortunately, Asylum Research uses Igor Pro 6! Thus, we need to try to
get this compiled for Igor Pro 6 as well. Looking at the differences
between 6, 7, and 8, it appears that there are no *major* changes at the
programmer level (i.e., the major changes are mostly abstracted away
from the Igor programmer). Of note: object names are limited to 31
bytes.

Thus, in this CL we try to change the major limitations:
- We rename the variables we saw that were larger than 31
characters (there are probably more!);
- We modify the IgorVersion pragma for associated ipfs to be 6 instead
of 8;
- We change the minimum igor version in the CMakeLists to 6, and modify
the igor paths accordingly.