feat: initial updates to readme to clarify
Here, we expan upon the original compilation instructions (and
dependencies list).

-The original version did not explicit that there
were multiple submodule dependencies, which needed to be initialized and
updated in order to compile.
- It also did not mention the requirement to specify where XOPSupport
is.

This is a first pass, as we have not updated the CMakeLists to allow
XOPSupport to be simply pointed to (or done the same for the launcher
batch script).