fix: update zmq folder (cmake) for output changes
The way the CMakeLists is set up, the headers (and libraries themselves)
are copied to a particular install folder, and linked from there. When I
changed the output folder structure, I forgot to do so for the reference
to the libzmq folder in the cmake. The result: when trying to compile,
it did not know where the libzmq headers were!

Fixed (I presume) in this CL, by updating the path it expects to include ${igorSubdir}.