refactor: rename igorSubdir to igorVersion (cmake)
I realized in my prior CL that igorSubdir is a poor/confusing name. Even
igorVersion, to the same point, is confusing. Really, we are dealing
with the XOP toolkits: XOP Toolkit 7 supports Igors 6 and 7, while XOP
Toolkit 8 supports Igors 8+.

I am sticking with the existing nomenclature because (a) I am somewhat
lazy, and (b) the odd person that needs Igor 6 support will read the
readme and follow the directions.