refactor: misc changes following merge review
- We are removing the ZMQ -> ZeroMQ renaming for all .cpp/.h
files (since not necessary).
- For ipf files, we are double-defining ZMQ_XXX and ZeroMQ_XXX variables,
for prior compatibility.
- We reverted the changes creating igor6/igor8 subdirectories in the
output folder. Now it all just goes into output.
- We are using a cmake variable, XOP_MINIMUM_VERSION to define the error
message if your minimum version is not supported. So now
ZeroMQWinCustom.rc and ZeroMQ.r are .in files and get configured via cmake.
- We are using a cmake variable, DECLARE_ZEROMQ_CONSTANTS_LEGACY, to
declare 'ZeroMQ' constants for the .ipf files in addition to the new ZMQ
format. This is done for prior versions. Note that this declaration only
happens if DECLARE_ZEROMQ_CONSTANTS_LEGACY is true, which occurs under
standard compilation (i.e. not using IgorPro 6.37).
- To choose between IGOR6 an IGOR8, you now simply override
XOP_MINIMUM_VERSION to be 637.