refactor: update readme + revert output structure
- We made some changes to how this new logic functions, but had not
updated the readme to indicate them. Doing so now.
- We revert all output folders to the original. In prior changelists, I
tried to 're-organize' the output structure, as I had introduced
individual directories for igor6 and igor8 builds. Now that we have
removed these (as including them did not make sense), I am returning to
the original output structure in main.
- We remove a previously added, now unused, variable from the CMakeLists.
- We document the purpose of MSVC_RUNTIME_DYNAMIC in the CMakeLists (I
found it a bit unclear).