feat: remove CMAKE_BUILD_TYPE from cmake build
While working on something else, I realized that the CMakeLists and
associated readmes use a mix of $<CONFIG> and ${CMAKE_BUILD_TYPE} for
the configuration. Since the supported generators are xcode for OSX and
Visual Studio for Mac, it may make sense to replace associated
CMAKE_BUILD_TYPE logic with the appropriate $<CONFIG> generator
expressions. This changelist does that.

NOTE: By removing ${CMAKE_BUILD_TYPE}, this changelist effectively
breaks the possibility of building using UNIX Makefiles, which *could*
in principle be a desired generator on OSX. If so, this changelist
should not be accepted!

In terms of changes, we primarily simply replace the cmake variable with
the generator expression, where appropriate. The noteworthy changes are:

- Tied to InstallRequiredSystemVariables, in src/CMakeLists.txt:
Since this function does not support generator expressions, we run it
2x, setting a local variable with the runtime libraries between them.
Once done, we manually invoke install(PROGRAMS) calls for Release and
Debug configurations.
- Since this CMake seems to be based around only Debug and Release
  builds, we limit the configurations to those 2 options.