feat: update to support Igor Pro 6
Making changes to the code to support Igor Pro 6 (the Igor Pro version
used by Asylum Research). The existing version of ZeroMQ-XOP supports
Igor versions > 8, presumably because it was compiled with the XOP
Toolkit 8 (and had no reason to support earlier versions).

Unfortunately, Asylum Research uses Igor Pro 6! Thus, we need to try to
get this compiled for Igor Pro 6 as well. Looking at the differences
between 6, 7, and 8, it appears that there are no *major* changes at the
programmer level (i.e., the major changes are mostly abstracted away
from the Igor programmer). Of note: object names are limited to 31
bytes.

Thus, in this CL we try to change the major limitations:
- We rename the variables we saw that were larger than 31
characters (there are probably more!);
- We modify the IgorVersion pragma for associated ipfs to be 6.20 instead
of 8. NOTE: we keep the minimum version for the tests to be 8, since
they assume an internal Igor method, JSONSimple, which does
not exist before this. Thus, the UTs *cannot* be run if compiling for Igor6/7.
- We change the minimum igor version in the CMakeLists to 6.20, and modify
the igor paths accordingly.
- We add FV_NORETURN_TYPE definition if compiling with XOP Toolkit 7. Even
though it is not used, its existence in the files causes a compile error
if not defined.
- We add an option to the CMakeLists, to choose between Igor6/7 and Igor8.
- We update the readme to explain this Igor6/7 support.
- We add a comment in the CMakeLists about OSX support with Igor as a
note (for future reference).