feat: changing output structure
- We modify where the executables are placed in the output folder on
install. Now, they are placed in output/${igorVersion}, where this may be
Igor6 or Igor8.
- We update the readme to indicate these changes.

refactor: moving output to output/igor8

We are doing this because we will be supporting igor6/7. Thus, igor6/7
builds will be located in output/igor6.

refactor: rename igorSubdir to igorVersion (cmake)

I realized in my prior CL that igorSubdir is a poor/confusing name. Even
igorVersion, to the same point, is confusing. Really, we are dealing
with the XOP toolkits: XOP Toolkit 7 supports Igors 6 and 7, while XOP
Toolkit 8 supports Igors 8+.

I am sticking with the existing nomenclature because (a) I am somewhat
lazy, and (b) the odd person that needs Igor 6 support will read the
readme and follow the directions.

refactor: move libzmq files to output/libzmq

It doesn't make sense for them to be in igor8/igor6, as they are the
same regardless!

Also, we update min igor6 version to 6.37. It was 6.20, according to the
documentation. t-b rightfully suggested that we force any user using
this experimental build to be on the latest version of Igor6.
Effectively, this is putting them on the 'old bleeding edge' (bleeding
edge of Igor 6, which is ooold).