fix: avoid apple compile error
indicate ${RESOURCES} is generated, to hopefully avoid a 'file not
found' error when prepping the build.

Note the description of what doing this does (CMake manual):
"Tells the internal CMake engine that a source file is generated by an outside process such as another build step, or the execution of CMake itself. This information is then used to exempt the file from any existence or validity checks."